[Firebase] Move sharing logic files into a directory
