+++ firebase/functions/src/share/appleMaps.ts
