+++ firebase/functions/src/share/appleMusic.ts
