+++ firebase/functions/src/share/inputData.ts
