+++ firebase/functions/src/share/item.ts
