+++ firebase/functions/src/share/normalizedData.ts
