+++ firebase/functions/src/share/util.ts
